+++ config/autoload/ems.global.php
@@ -1,7 +1,0 @@
-<?php
-
-return [
-    'dot_ems' => [
-        'default_adapter' => 'database',
-    ]
-];