+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## 0.2.0 - 2017-03-22
+
+ZE2 migration
+
+### Changed
+* updated factories to PSR11 container and middleware to PSR15
+* refactored folder structure to follow modular application design a la ZE2
+* programmatic pipeline and routes
+* restructured application to match ZE2 structure
+
+### Added
+* Nothing
+
+### Deprecated
+* Nothing
+
+### Removed
+* Nothing
+
+### Fixed
+* Nothing
+
+
 ## 0.1.1 - 2017-03-12
 
 ### Added